feat: 렌더링 최적화 - 심화
