+++ packages/assignment/src/context/NotificationContext.ts
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+import type { NotificationContextType } from "../types";
+
+const NotificationContext = createContext<NotificationContextType | undefined>(undefined);
+
+export default NotificationContext;
