+++ packages/assignment/src/context/ThemeContext.ts
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+import type { ThemeContextType } from "../types";
+
+const ThemeContext = createContext<ThemeContextType | undefined>(undefined);
+
+export default ThemeContext;
