+++ packages/assignment/src/context/UserContext.ts
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+import type { UserContextType } from "../types";
+
+const UserContext = createContext<UserContextType | undefined>(undefined);
+
+export default UserContext;
