+++ packages/assignment/src/context/index.ts
@@ -1,0 +1,5 @@
+import NotificationContext from "./NotificationContext";
+import ThemeContext from "./ThemeContext";
+import UserContext from "./UserContext";
+
+export { NotificationContext, ThemeContext, UserContext };
