+++ packages/assignment/src/hooks/useUser.ts
@@ -1,0 +1,10 @@
+import { useContext } from "react";
+import { UserContext } from "../context";
+
+export const useUser = () => {
+    const context = useContext(UserContext);
+
+    if (!context) throw new Error('useUser must be used within a UserProvider');
+    
+    return context;
+};
